+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Use Inter as default font
 - Add forgot password info during unlock
 - Export transactions CSV: add historical value and historical value currency columns.
+- Persist account receive script type
 
 ## v4.50.1
 - Fix a bug that would delay showing watch-only accounts.